add chrome browser-use selector to onboarding wizard, add pfp selection to wizard, cleanup
